JavaScript Long/Int broken time logging
Former-commit-id: cfd4a394be1030e6cdb4e4dccda1554f802c00a5

+++ Jmol/src/org/jmol/util/Logger.java
@@ -301,22 +301,22 @@ public final class Logger {
     }
   }
 
-  static Map<String, Integer>htTiming = new Hashtable<String, Integer>();
+  static Map<String, Long>htTiming = new Hashtable<String, Long>();
   public static void startTimer(String msg) {
     if (msg != null)
-      htTiming.put(msg, Integer.valueOf((int) System.currentTimeMillis()));
+      htTiming.put(msg, Long.valueOf(System.currentTimeMillis()));
   }
 
-  public static String getTimerMsg(String msg, long time) {
+  public static String getTimerMsg(String msg, int time) {
     if (time == 0)
       time = getTimeFrom(msg);
     return "Time for " + msg + ": " + (time) + " ms";
   }
   
   private static int getTimeFrom(String msg) {
-    Integer t;
-    return (msg == null || (t = htTiming.get(msg)) == null ? -1 : (int) (System
-        .currentTimeMillis() - t.intValue()));
+    Long t;
+    return (int) (msg == null || (t = htTiming.get(msg)) == null ? -1 : System
+        .currentTimeMillis() - t.longValue());
   }
 
   public static int checkTimer(String msg, boolean andReset) {
